# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'AZE FARM API'
  version: 1.0.0
  description: 'API pour la gestion des produits agricoles'
servers:
  -
    url: 'http://localhost:3000/api'
    description: 'Serveur de développement'
paths:
...
  /api/users/avatar:
    put:
      summary: 'Mettre à jour l''avatar de l''utilisateur'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Met à jour uniquement l''avatar de l''utilisateur connecté'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
                  description: 'Nouvelle photo de profil (formats jpg, png, webp, max 5MB)'
      responses:
        '200':
          description: 'Avatar mis à jour avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Avatar updated successfully'}
                  data: {type: object, properties: {user: {type: object, properties: {user_id: {type: string}, email: {type: string}, fullname: {type: string}, avatar_url: {type: string, example: 'https://minio.example.com/avatar/user123.jpg'}}}}}
        '400':
          description: 'Fichier avatar manquant ou invalide'
        '401':
          description: 'Non authentifié'
        '403':
          description: 'Compte non vérifié ou inactif'
        '413':
          description: 'Fichier trop volumineux'
        '415':
          description: 'Format de fichier non supporté'
        '500':
          description: 'Erreur serveur'
  /api/users/search:
    get:
      summary: 'Rechercher des utilisateurs'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Recherche des utilisateurs par nom ou email'
      parameters:
        -
          in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 50
          description: 'Terme de recherche (nom ou email)'
        -
          in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: 'Nombre maximum de résultats'
        -
          in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 'Numéro de page'
      responses:
        '200':
          description: 'Utilisateurs trouvés avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Utilisateurs trouvés'}
                  data: {type: object, properties: {users: {type: array, items: {type: object, properties: {user_id: {type: string}, email: {type: string}, fullname: {type: string}, avatar_url: {type: string}, role: {type: string, enum: [user, farmer, admin]}, is_verified: {type: boolean}, is_active: {type: boolean}, createdAt: {type: string, format: date-time}}}}, pagination: {type: object, properties: {page: {type: integer}, limit: {type: integer}, totalPages: {type: integer}, totalItems: {type: integer}}}}}
        '400':
          description: 'Paramètres de recherche invalides'
        '401':
          description: 'Non authentifié'
        '500':
          description: 'Erreur serveur'
  /api/users:
    get:
      summary: 'Lister tous les utilisateurs'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Récupère la liste de tous les utilisateurs avec pagination'
      parameters:
        -
          in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 'Numéro de page'
        -
          in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: 'Nombre d''utilisateurs par page'
        -
          in: query
          name: role
          schema:
            type: string
            enum:
              - user
              - farmer
              - admin
          description: 'Filtrer par rôle'
        -
          in: query
          name: is_verified
          schema:
            type: boolean
          description: 'Filtrer par statut de vérification'
        -
          in: query
          name: is_active
          schema:
            type: boolean
          description: 'Filtrer par statut actif'
      responses:
        '200':
          description: 'Utilisateurs récupérés avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Utilisateurs récupérés'}
                  data: {type: object, properties: {users: {type: array, items: {type: object, properties: {user_id: {type: string}, email: {type: string}, fullname: {type: string}, avatar_url: {type: string}, role: {type: string, enum: [user, farmer, admin]}, is_verified: {type: boolean}, is_active: {type: boolean}, createdAt: {type: string, format: date-time}, last_login: {type: string, format: date-time}}}}, pagination: {type: object, properties: {page: {type: integer}, limit: {type: integer}, totalPages: {type: integer}, totalItems: {type: integer}}}}}
        '401':
          description: 'Non authentifié'
        '500':
          description: 'Erreur serveur'
  /api/users/export:
    get:
      summary: 'Exporter les utilisateurs en CSV'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Exporte la liste de tous les utilisateurs au format CSV (réservé aux administrateurs)'
      responses:
        '200':
          description: 'Fichier CSV généré avec succès'
          content:
            text/csv:
              schema:
                type: string
                description: 'Contenu du fichier CSV'
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'attachment; filename="users_export.csv"'
            Content-Type:
              schema:
                type: string
              example: text/csv
        '401':
          description: 'Non authentifié'
        '403':
          description: 'Droits administrateur requis'
        '500':
          description: 'Erreur lors de l''export'
  '/api/users/{user_id}':
    get:
      summary: 'Récupérer les détails d''un utilisateur'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Récupère les informations complètes d''un utilisateur spécifique'
      parameters:
        -
          in: path
          name: user_id
          required: true
          schema:
            type: string
          description: 'ID de l''utilisateur'
      responses:
        '200':
          description: 'Utilisateur trouvé avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Utilisateur trouvé'}
                  data: {type: object, properties: {user: {type: object, properties: {user_id: {type: string}, email: {type: string}, fullname: {type: string}, avatar_url: {type: string}, role: {type: string, enum: [user, farmer, admin]}, is_verified: {type: boolean}, is_active: {type: boolean}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}, last_login: {type: string, format: date-time}}}}}
        '400':
          description: 'ID utilisateur invalide'
        '401':
          description: 'Non authentifié'
        '404':
          description: 'Utilisateur non trouvé'
        '500':
          description: 'Erreur serveur'
    delete:
      summary: 'Supprimer un utilisateur (soft delete)'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Marque un utilisateur comme supprimé (soft delete) - réservée aux administrateurs'
      parameters:
        -
          in: path
          name: user_id
          required: true
          schema:
            type: string
          description: 'ID de l''utilisateur à supprimer'
      responses:
        '200':
          description: 'Utilisateur supprimé avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Utilisateur supprimé avec succès'}
        '400':
          description: 'ID utilisateur invalide'
        '401':
          description: 'Non authentifié'
        '403':
          description: 'Droits administrateur requis'
        '404':
          description: 'Utilisateur non trouvé'
        '500':
          description: 'Erreur serveur'
  '/api/users/{user_id}/role':
    put:
      summary: 'Mettre à jour le rôle d''un utilisateur'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Change le rôle d''un utilisateur (réservé aux administrateurs)'
      parameters:
        -
          in: path
          name: user_id
          required: true
          schema:
            type: string
          description: 'ID de l''utilisateur'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [user, farmer, admin]
                  description: 'Nouveau rôle de l''utilisateur'
      responses:
        '200':
          description: 'Rôle mis à jour avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Rôle mis à jour avec succès'}
                  data: {type: object, properties: {user: {type: object, properties: {user_id: {type: string}, email: {type: string}, fullname: {type: string}, role: {type: string, enum: [user, farmer, admin]}, updatedAt: {type: string, format: date-time}}}}}
        '400':
          description: 'Données invalides ou rôle non autorisé'
        '401':
          description: 'Non authentifié'
        '403':
          description: 'Droits administrateur requis'
        '404':
          description: 'Utilisateur non trouvé'
        '500':
          description: 'Erreur serveur'
  '/api/users/{user_id}/restore':
    post:
      summary: 'Restaurer un utilisateur supprimé'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Restaure un utilisateur précédemment supprimé (soft delete) - réservée aux administrateurs'
      parameters:
        -
          in: path
          name: user_id
          required: true
          schema:
            type: string
          description: 'ID de l''utilisateur à restaurer'
      responses:
        '200':
          description: 'Utilisateur restauré avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Utilisateur restauré avec succès'}
                  data: {type: object, properties: {user: {type: object, properties: {user_id: {type: string}, email: {type: string}, fullname: {type: string}, is_active: {type: boolean, example: true}, restoredAt: {type: string, format: date-time}}}}}
        '400':
          description: 'ID utilisateur invalide'
        '401':
          description: 'Non authentifié'
        '403':
          description: 'Droits administrateur requis'
        '404':
          description: 'Utilisateur non trouvé ou pas supprimé'
        '500':
          description: 'Erreur serveur'
...